Initial Proof of Concept Code
This is the basic POC code to prove how this can be accomplished.
At this point, stabalization, testing, and some refactoring is still needed.
Also a sane configuration needs to be set in order to allow for multiple
database backends.  I'm also contemplating adding in a way to link this
into the ticketing system.